Feature/vincent fix 2

+++ rigsg/entity-model/src/main/java/com/nus/project/capstone/model/entity/idc/PresentationResponse.java
@@ -1,7 +1,6 @@
 package com.nus.project.capstone.model.entity.idc;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.nus.project.capstone.model.persistence.idc.IdcTeamJpaEntities;
 import com.nus.project.capstone.model.persistence.idc.PresentationJpaEntities;
 import lombok.Builder;
 import lombok.Data;
@@ -16,7 +15,7 @@ public class PresentationResponse {
 
     private UUID id;
     private String stage;
-    private Integer score;
+    private Float score;
     private String venue;
     private Instant dateTime;
